% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfbsgdp.R
\name{cdfbsgdp}
\alias{cdfbsgdp}
\title{Cumulative distribution function of the Birnbaum-Saunders-Generalized Pareto distribution}
\usage{
cdfbsgdp(par, x)
}
\arguments{
\item{par}{parameter vector of the Birnbaum-Saunders-Generalized Pareto distribution.
First parameter is the shape, second parameter is the scale parameter. Third parameter is the lower bound parameter.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the cdf of the Birnbaum-Saunders-Generalized Pareto  distribution
}
\description{
Cumulative distribution function of the Birnbaum-Saunders-Generalized Pareto distribution
}
\examples{
cdfbsgdp(c(0.5,2,0.5),3)
}
\references{
Altun, E., Ozel, G. A novel approach to probabilistic hazard assessment: BSGPD model. (Under Review)
}
