% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfeexp.R
\name{cdfeexp}
\alias{cdfeexp}
\title{Cumulative distribution function of the exponentiated exponential distribution}
\usage{
cdfeexp(par, x)
}
\arguments{
\item{par}{parameter vector of the exponentiated exponential distribution.
First parameter is the shape, second is the scale parameter.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the exponentiated exponential distribution
}
\description{
Cumulative distribution function of the exponentiated exponential distribution
}
\examples{
cdfeexp(c(0.5,0.3),2)
}
\references{
Gupta, R. D., & Kundu, D. (1999). Theory & methods: Generalized exponential distributions. Australian & New Zealand Journal of Statistics, 41(2), 173-188.
}
