% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfer.R
\name{cdfer}
\alias{cdfer}
\title{Cumulative distribution function of the exponentiated Rayleigh  distribution}
\usage{
cdfer(par, x)
}
\arguments{
\item{par}{parameter vector of the exponentiated Rayleigh  distribution.
First parameter is the scale, second is the shape parameter.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the exponentiated Rayleigh  distribution
}
\description{
Cumulative distribution function of the exponentiated Rayleigh  distribution
}
\examples{
cdfer(c(0.5,0.3),2)
}
\references{
Vodă, V. G. (1976). Inferential procedures on a generalized Rayleigh variate. I. Aplikace matematiky, 21(6), 395-412.
}
