% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfew.R
\name{cdfew}
\alias{cdfew}
\title{Cumulative distribution function of the exponentiated Weibull distribution}
\usage{
cdfew(par, x)
}
\arguments{
\item{par}{parameter vector of the exponentiated Weibull distribution.
First parameter is the shape, second is the scale parameter and third parameter is shape parameter.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the exponentiated Weibull distribution
}
\description{
Cumulative distribution function of the exponentiated Weibull distribution
}
\examples{
cdfew(c(0.5,0.3,0.6),2)
}
\references{
Mudholkar, G. S., & Srivastava, D. K. (1993). Exponentiated Weibull family for analyzing bathtub failure-rate data. IEEE transactions on reliability, 42(2), 299-302.
}
