% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfinvgamma.R
\name{cdfinvgamma}
\alias{cdfinvgamma}
\title{Cumulative distribution function of the inverse gamma distribution}
\usage{
cdfinvgamma(par, x)
}
\arguments{
\item{par}{parameter vector of the inverse gamma distribution.
First parameter is the shape, second is the rate parameter.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the inverse gamma distribution
}
\description{
Cumulative distribution function of the inverse gamma distribution
}
\examples{
cdfinvgamma(c(2,5,3),3)
}
\references{
Cook, J. D. (2008). Inverse gamma distribution. online: http://www. johndcook. com/inverse gamma. pdf, Tech. Rep.
}
