% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfiweibull.R
\name{cdfiwweibull}
\alias{cdfiwweibull}
\title{Cumulative distribution function of the inverse Weibull distribution}
\usage{
cdfiwweibull(par, x)
}
\arguments{
\item{par}{parameter vector of the inverse Weibull distribution.
First parameter is the shape and second is the scale parameter}

\item{x}{vector of quantiles}
}
\value{
return the value of the cdf of the inverse Weibull distribution
}
\description{
Cumulative distribution function of the inverse Weibull distribution
}
\examples{
cdfiwweibull(c(2,3),5)
}
\references{
Mudholkar, G. S., & Kollia, G. D. (1994). Generalized Weibull family: a structural analysis. Communications in statistics-theory and methods, 23(4), 1149-1171.
}
