% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdflevy.R
\name{cdflevy}
\alias{cdflevy}
\title{Cumulative distribution function of the Levy  distribution}
\usage{
cdflevy(par, x)
}
\arguments{
\item{par}{parameter vector of the Levy  distribution.
First parameter is the location, second is the scale parameter.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the Levy  distribution
}
\description{
Cumulative distribution function of the Levy  distribution
}
\examples{
cdflevy(c(0.5,0.3),2)
}
\references{
Nolan, J. P. (2003). Modeling financial data with stable distributions. In Handbook of heavy tailed distributions in finance (pp. 105-130). North-Holland.
}
