% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdflnormal.R
\name{cdflnormal}
\alias{cdflnormal}
\title{Cumulative distribution function of the log-normal distribution}
\usage{
cdflnormal(par, x)
}
\arguments{
\item{par}{parameter vector of the log-normal distribution.
First parameter is the shape and second is the scale parameter}

\item{x}{vector of quantiles}
}
\value{
return the value of the cdf of the log-normal distribution
}
\description{
Cumulative distribution function of the log-normal distribution
}
\examples{
cdflnormal(c(2,3),5)
}
\references{
Heyde, C. C. (1963). On a property of the lognormal distribution. Journal of the Royal Statistical Society: Series B (Methodological), 25(2), 392-393.
}
