% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfpareto.R
\name{cdfpareto}
\alias{cdfpareto}
\title{Cumulative distribution function of the Pareto distribution}
\usage{
cdfpareto(par, x)
}
\arguments{
\item{par}{parameter vector of the Pareto distribution.
First parameter is the shape and second is the scale parameter}

\item{x}{vector of quantiles}
}
\value{
return the value of the cdf of the Pareto distribution
}
\description{
Cumulative distribution function of the Pareto distribution
}
\examples{
cdfpareto(c(2,5),2)
}
\references{
Arnold, B. C. (1983). Pareto Distributions, International Cooperative Publishing House.
}
