% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfrayleigh.R
\name{cdfrayleigh}
\alias{cdfrayleigh}
\title{Cumulative distribution function of the Rayleigh distribution}
\usage{
cdfrayleigh(par, x)
}
\arguments{
\item{par}{scale parameter vector of the Rayleigh distribution.}

\item{x}{vector of quantiles}
}
\value{
return the value of the cdf of the Rayleigh distribution
}
\description{
Cumulative distribution function of the Rayleigh distribution
}
\examples{
cdfrayleigh(c(2),5)
}
\references{
Siddiqui, M. M. (1964). Statistical inference for Rayleigh distributions. Journal of Research of the National Bureau of Standards, Sec. D, 68(9), 1005-1010.
}
