% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_earthquake_6_6.5.R
\docType{data}
\name{data_earthquake_6_6.5}
\alias{data_earthquake_6_6.5}
\title{Earthquake dataset}
\format{
A numeric vector
}
\usage{
data_earthquake_6_6.5
}
\description{
The elapsed time (year) between the earthquakes with 6 and 6.5 magnitudes in Turkey occured between the years of 1990-2021
}
\keyword{datasets}
