% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_earthquake_7.R
\docType{data}
\name{data_earthquake_7}
\alias{data_earthquake_7}
\title{Earthquake dataset}
\format{
A numeric vector
}
\usage{
data_earthquake_7
}
\description{
The elapsed time (year) between the earthquakes having the magnitudes higher than 7 in Turkey occured between the years of 1990-2021
}
\keyword{datasets}
