% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitexpexp.R
\name{fitexpexp}
\alias{fitexpexp}
\title{Fitting the exponentiated exponential distribution}
\usage{
fitexpexp(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the exponentiated exponential distribution
}
\examples{
data=rexpexp(500,2,3)
fitexpexp(starts =c(2,2),data=data)
}
