% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitexprayleigh.R
\name{fitexprayleigh}
\alias{fitexprayleigh}
\title{Fitting the exponentiated exponentiated Rayleigh distribution}
\usage{
fitexprayleigh(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the exponentiated exponentiated Rayleigh distribution
}
\examples{
data=rexprayleigh(500,2,3)
fitexprayleigh(starts =c(2,2),data=data)
}
