% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitexpweibull.R
\name{fitexpweibull}
\alias{fitexpweibull}
\title{Fitting the exponentiated Weibull distribution}
\usage{
fitexpweibull(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the exponentiated Weibull distribution
}
\examples{
data=rexpweibull(500,2,3,5)
fitexpweibull(starts =c(2,2,2),data=data)
}
