% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgamma.R
\name{fitgamma}
\alias{fitgamma}
\title{Fitting the gamma distribution}
\usage{
fitgamma(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the gamma distribution
}
\examples{
datagamma=rgamma(500,2,2)
fitgamma(starts =c(2,2),data=datagamma)
}
