% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitggamma.R
\name{fitggamma}
\alias{fitggamma}
\title{Fitting the generalized gamma distribution}
\usage{
fitggamma(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the generalized gamma distribution
}
\examples{
library(rmutil)
data=rggamma(500,2,2,2)
fitggamma(starts =c(1,1,1),data=data)
}
