% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgumbel.R
\name{fitgumbel}
\alias{fitgumbel}
\title{Fitting the Gumbel distribution}
\usage{
fitgumbel(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the Gumbel distribution
}
\examples{
library(VGAM)
data=rgumbel(500,2,0.5)
fitgumbel(starts =c(2,2),data=data)
}
