% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitinvgamma.R
\name{fitinvgamma}
\alias{fitinvgamma}
\title{Fitting the inverse gamma distribution}
\usage{
fitinvgamma(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the inverse gamma distribution
}
\examples{
library(invgamma)
data=rinvgamma(500,2,0.5)
fitinvgamma(starts =c(2,2),data=data)
}
