% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitiwebull.R
\name{fitiweibull}
\alias{fitiweibull}
\title{Fitting the gamma distribution}
\usage{
fitiweibull(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the gamma distribution
}
\examples{
set.seed(7)
data=rgamma(500,shape=1,scale=1)
fitiweibull(starts =c(0.5,0.5),data=data)
}
