% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlevy.R
\name{fitlevy}
\alias{fitlevy}
\title{Fitting the Levy distribution}
\usage{
fitlevy(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the Levy distribution
}
\examples{
library(VGAM)
data=ERPeq::rlevy(100,2,0.1)
fitlevy(starts =c(0.1),data=data)
}
