% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlnormal.R
\name{fitlnormal}
\alias{fitlnormal}
\title{Fitting the log-normal distribution}
\usage{
fitlnormal(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the log-normal distribution
}
\examples{
data=rlnorm(500,2,0.5)
fitlnormal(starts =c(2,2),data=data)
}
