% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitrayleigh.R
\name{fitrayleigh}
\alias{fitrayleigh}
\title{Fitting the Rayleigh  distribution}
\usage{
fitrayleigh(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the Rayleigh  distribution
}
\examples{
library(VGAM)
data=rrayleigh(500,2)
fitrayleigh(starts =c(2),data=data)
}
