% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitweibull.R
\name{fitweibull}
\alias{fitweibull}
\title{Fitting the Weibull distribution}
\usage{
fitweibull(starts, data)
}
\arguments{
\item{starts}{A vector defining the starting values for the Nelder-Mead algorithm.}

\item{data}{A vector containing the observations}
}
\value{
List the estimated parameters of the distribution with standard errors and goodness-of-fit statistics.
}
\description{
Fitting the Weibull distribution
}
\examples{
dataweibull=rweibull(500,2,2)
fitweibull(starts =c(2,2),data=dataweibull)
}
