% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfbsgdp.R
\name{pdfbsgdp}
\alias{pdfbsgdp}
\title{Probability density function of the Birnbaum-Saunders-Generalized Pareto distribution}
\usage{
pdfbsgdp(par, x)
}
\arguments{
\item{par}{parameter vector of the Birnbaum-Saunders-Generalized Pareto distribution.
First parameter is the shape, second parameter is the scale parameter. Third parameter is the lower bound parameter.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the Birnbaum-Saunders-Generalized Pareto  distribution.
}
\description{
Probability density function of the Birnbaum-Saunders-Generalized Pareto distribution
}
\examples{
pdfbsgdp(c(2,0.5,0.5),1)
}
\references{
Altun, E., Ozel, G. A novel approach to probabilistic hazard assessment: BSGPD model. (Under Review)
}
