% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfgamma.R
\name{pdfgamma}
\alias{pdfgamma}
\title{Probability density function of the Gamma distribution}
\usage{
pdfgamma(par, x)
}
\arguments{
\item{par}{parameter vector of the gamma distribution.
First parameter is the shape and second is the scale parameter}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the gamma distribution
}
\description{
Probability density function of the Gamma distribution
}
\examples{
pdfgamma(c(2,3),5)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 21. Wiley, New York.
}
