% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfggamma.R
\name{pdfggamma}
\alias{pdfggamma}
\title{Probability density function of the generalized gamma distribution}
\usage{
pdfggamma(par, x)
}
\arguments{
\item{par}{parameter vector of the generalized gamma distribution.
First parameter is the dispersion, second is the location parameter and third is the family parameter.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the generalized gamma distribution
}
\description{
Probability density function of the generalized gamma distribution
}
\examples{
pdfggamma(c(2,5,3),3)
}
\references{
Stacy, E. W. (1962). A generalization of the gamma distribution. The Annals of mathematical statistics, 1187-1192.
}
