% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfgumbel.R
\name{pdfgumbel}
\alias{pdfgumbel}
\title{Probability density function of the gumbel distribution}
\usage{
pdfgumbel(par, x)
}
\arguments{
\item{par}{parameter vector of the gumbel distribution.
First parameter is the location, second is the scale parameter.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the gumbel distribution
}
\description{
Probability density function of the gumbel distribution
}
\examples{
pdfgumbel(c(0.5,0.3),2)
}
\references{
Gumbel, E. J. (1941). The return period of flood flows. The annals of mathematical statistics, 12(2), 163-190.
}
