% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfiweibull.R
\name{pdfiweibull}
\alias{pdfiweibull}
\title{Probability density function of the inverse Weibull distribution}
\usage{
pdfiweibull(par, x)
}
\arguments{
\item{par}{parameter vector of the inverse Weibull distribution.
First parameter is the shape and second is the scale parameter}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the inverse Weibull distribution
}
\description{
Probability density function of the inverse Weibull distribution
}
\examples{
pdfiweibull(c(2,3),5)
}
\references{
Mudholkar, G. S., & Kollia, G. D. (1994). Generalized Weibull family: a structural analysis. Communications in statistics-theory and methods, 23(4), 1149-1171.
}
