% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfpareto.R
\name{pdfpareto}
\alias{pdfpareto}
\title{Probability density function of the Pareto distribution}
\usage{
pdfpareto(par, x)
}
\arguments{
\item{par}{parameter vector of the Pareto distribution.
First parameter is the scale and second is the shape parameter}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the Pareto distribution
}
\description{
Probability density function of the Pareto distribution
}
\examples{
pdfpareto(c(2,5),3)
}
\references{
Arnold, B. C. (1983). Pareto Distributions, International Cooperative Publishing House.
}
