% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfrayleigh.R
\name{pdfrayleigh}
\alias{pdfrayleigh}
\title{Probability density function of the Rayleigh distribution}
\usage{
pdfrayleigh(par, x)
}
\arguments{
\item{par}{scale parameter vector of the Rayleigh distribution.}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the Rayleigh distribution
}
\description{
Probability density function of the Rayleigh distribution
}
\examples{
pdfrayleigh(c(2),5)
}
\references{
Siddiqui, M. M. (1964). Statistical inference for Rayleigh distributions. Journal of Research of the National Bureau of Standards, Sec. D, 68(9), 1005-1010.
}
