% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfweibull.R
\name{pdfweibull}
\alias{pdfweibull}
\title{Probability density function of the Weibull distribution}
\usage{
pdfweibull(par, x)
}
\arguments{
\item{par}{parameter vector of the weibull distribution.
First parameter is the shape and second is the scale parameter}

\item{x}{vector of observations or single value}
}
\value{
return the value of the pdf of the weibull distribution
}
\description{
Probability density function of the Weibull distribution
}
\examples{
pdfweibull(c(2,3),5)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 21. Wiley, New York.
}
