% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbsgpd.R
\name{rbsgpd}
\alias{rbsgpd}
\title{Generate random observations from Birnbaum-Saunders-Generalized Pareto distribution}
\usage{
rbsgpd(n, beta, alpha, gamma)
}
\arguments{
\item{n}{number of observations to be generated from the Birnbaum-Saunders-Generalized Pareto}

\item{beta}{lower bound parameter of the}

\item{alpha}{scale parameter of the Birnbaum-Saunders-Generalized Pareto distribution}

\item{gamma}{shape parameter of the Birnbaum-Saunders-Generalized Pareto distribution}
}
\value{
return the random sample generated from scale parameter of the Birnbaum-Saunders-Generalized Pareto distribution distribution
}
\description{
Generate random observations from Birnbaum-Saunders-Generalized Pareto distribution
}
\examples{
rbsgpd(100,2,3,5)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 21. Wiley, New York.
}
