% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexpexp.R
\name{rexpexp}
\alias{rexpexp}
\title{Generate random observations from exponentiated exponential distribution}
\usage{
rexpexp(n, alpha, lambda)
}
\arguments{
\item{n}{number of observations to be generated}

\item{alpha}{shape parameter of the exponentiated exponential distribution}

\item{lambda}{scale parameter of the exponentiated exponential distribution}
}
\value{
return the random sample generated from exponentiated exponential distribution
}
\description{
Generate random observations from exponentiated exponential distribution
}
\examples{
rexpexp(100,2,3)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 21. Wiley, New York.
}
