% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexprayleigh.R
\name{rexprayleigh}
\alias{rexprayleigh}
\title{Generate random observations from exponentiated Rayleigh distribution}
\usage{
rexprayleigh(n, alpha, beta)
}
\arguments{
\item{n}{number of observations to be generated}

\item{alpha}{shape parameter of the exponentiated Rayleigh distribution}

\item{beta}{scale parameter of the exponentiated Rayleigh distribution}
}
\value{
return the random sample generated from exponentiated exponential distribution
}
\description{
Generate random observations from exponentiated Rayleigh distribution
}
\examples{
rexprayleigh(100,2,3)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 21. Wiley, New York.
}
