% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexpweibull.R
\name{rexpweibull}
\alias{rexpweibull}
\title{Generate random observations from exponentiated Weibull distribution}
\usage{
rexpweibull(n, alpha, beta, theta)
}
\arguments{
\item{n}{number of observations to be generated}

\item{alpha}{shape parameter of the exponentiated Weibull distribution}

\item{beta}{scale parameter of the exponentiated Weibull distribution}

\item{theta}{shape parameter of the exponentiated Weibull distribution}
}
\value{
return the random sample generated from exponentiated Weibull distribution
}
\description{
Generate random observations from exponentiated Weibull distribution
}
\examples{
rexpweibull(100,2,3,2)
}
\references{
Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 21. Wiley, New York.
}
